destroyuser

+++ src/controllers/users/destroyUser.js
@@ -2,16 +2,21 @@ const httpStatus = require('http-status');
 const { dbModel } = require('../../db/sequelize');
 const User = dbModel.tblUsers;
 
-exports.destroyOneUser = (req, res) => {
-  User.findByPk(req.params.id).then(user => {
-    const userDeleted = user;
-    User.destroy({
-      where: { id: user.id }
-    })
-      .then(_ => {
-        const message = `L'utilisateur ${userDeleted.last_name} ${userDeleted.first_name} a bien été supprimé.`;
-        res.json({ message, data: userDeleted });
-        console.log(userDeleted.toJSON());
+exports.destroyOneUser = async (req, res) => {
+  if (req.params.id) {
+    try {
+      const deleteUser = await User.findByPk(req.params.id);
+      await User.destroy({
+        where: { id: deleteUser.id }
       });
-  });
+      const message = `L'utilisateur ${deleteUser.last_name} ${deleteUser.first_name} a bien été supprimé.`;
+      res.json({ message, data: deleteUser });
+      console.log(deleteUser.toJSON());
+    } catch (error) {
+      console.log(error.message);
+      return res
+        .status(httpStatus.BAD_REQUEST)
+        .json({ error: 'la requete a échouée' });
+    }
+  }
 };
